Fixes #5495
method `select` is overwritten to shift `this.active` from the `.all-categories` link to the clicked menu node. After opening a menu item `this.active` is filled with the `.all-categories` link and the `select` method does not expect that. It closes all menu items and skips the activation of the selected menu item.

A `collapseAll` event is triggered for mobile devices if an open menu item is clicked so it closes it again.
method `expand` is overwritten to skip expansion of the clicked menu node of the menu node was already open. This event would otherwise take place directly after the added `collapseAll`.

Although this is a working solution it does rewrite two major methods of the jQuery UI menu which I feel should not be necessary. So if someone would like to look into a different solution, be welcome.