Remove obsolete comment in catalog_category_view.xml
In catalog_category_view.xml there is an XML comment that tells developers they can change the pager increments using methods like "addPagerLimit". These settings can be found in the backend now and the method "addPagerLimit" is only to be found in a list of deprecated methods I think it is safe to assume this comment doesn't serve a purpose anymore.