Changed locator class name for ObjectManager
Changed the locator class name for the object manager to \Magento\Framework\App\ObjectManager which extends \Magento\Framework\ObjectManager\ObjectManager. In the constructor of \Magento\Framework\App\ObjectManager the object manager is already set into the static variable of the class, so there is no need to call it seperately.

If this pull request is denied, then the constructor function of \Magento\Framework\App\ObjectManager should be removed since it will not be called anywhere in the application.