Fix Swagger-generated operation parameter names for "body" parameters
These were previously all output as a string "$body". This improves Swagger-generated API clients and compatibility with swagger-codegen & janephp/openapi generator.

This caused the following issues:

- Variable variable output in janephp/openapi code generation
- Incorrect code output (multiple models called "Body") in janephp/openapi
- "Anonymous" class naming in swagger-codegen generated clients (approx 20 models called Body1, Body2, Body3 are output for a Magento Enterprise schema)

The new format is: operationNamePostBody. Operation Id is guaranteed to be unique under the Open API specification. Changing the body parameter in the schema should also not affect clients conforming to the specification.