Add --refresh-content-version-only option to deploy
With this option you can quickly force a browser cache refresh or
a CDN cache refresh without having to generate all static content.

This comes in handy especially when you have made no code changes,
but the CDN failed to update it's cache the last time you deployed
or when you have made a manual change in one of the static files
as a hotfix on your deployment server, awaiting a suitable moment
for deployment to finalize it.

This reopens #9338.