Cron uses the wrong timestamp method
Because the cron job uses a localized timestamp instead of a regular
timestamp, all times in the cron_schedule table are not the actual
times, but times with a double offset.

By using the regular timestamp, all times in cron_schedule are correct.
The cron job execution times are not affected by this change because
schedule generation and job execution depend on the same timestamp.
If both are in the same timezone all will go as intended.

Fixes #4237