Allow option disabling for optgroup binding
The optgroup Knockout binding formats its options before usage in the optionsAfterRender of the form/element/select.html knockout template. In this process all extra data of the options will be stripped and the optionsAfterRenderer cannot access the disabled value of any option. Because of this the delimiter option in the country_id select in checkout is choosable by the customer.