Fix overwrite default value image/file with NULL
If you set a default value for a config field type image/file and you save it. It won't set the default value in the database. Instead it will save NULL and so overwrite the default config.

With this change the default value will be set as value.

This is a replication of what used to be in of https://github.com/magento/magento2/pull/5085