Correct flat indexing with staging and > 500 cats.
Previously, the array was prefilled with entity_id values to ensure those categories
were indexed.  Unfortunately, when using row_id, it was using entity_ids.

When chunking at 500, this could cause an overlapping row to be processed twice (the
second time with missing non-static attributes), and cause a duplicate key error.