Fix #10682: Meta description and keywords transform to html entities
Changed the htmlentities function to the htmlspecialchars function. It's not necessary to do HTML encoding on all characters as described in #10682. Since the content is used within HTML tags, we still need to encode special HTML chars in order to make sure that chars like "'& etc, will not break the page.