Send order email for Braintree Paypal orders
Order emails are sent using observers on the
`sales_model_service_quote_submit_success` event, but only in the
`webapi_rest` and `webapi_soap` areas. When customers checkout with
Braintree Paypal the order placement occurs in the `frontend` area with
a request to `Magento\Braintree\Controller\Paypal\PlaceOrder`, so these
orders were not sent a new order email.

Fixes #13778
Fixes #12792