Disable "Back" button on the first step of the setup
When you enter the setup of Magento, the "Back" button is enabled
on the first step. When you click this, you go straight to the
final step of the installation, without entering any data in the
steps, so you can't install Magento at all.

With this commit the back button on the first step of the setup
is always disabled, since there's no previous step, except for the
page where you agree with the terms and conditions, which is not
part of the wizard.