[2.3-develop] [ForwardPort] Port of #15020 Update Gallery Template to handle boolean config Variables
Due to changes implemented in the resolution to #12285, boolean configuration variables are now properly typed booleans, instead of the strings "true" and "false".

Without this fix applied, config vals that were true were being output in the gallery template javascript as : 1 and values that were false were being output as :

This causes javascript errors for any item that is set to false.