Make it possible to disable cross-sell on cart page
To be able to select whether to have the cross-sell active on the cart
page an admin configuration is created under Stores > Configuration >
Sales > Checkout > Shopping Cart.

The default for this setting is set to Yes (Show cross-sell items). This
is for backwards compatibility and not have webshop owners accidentally
not showing these items.

Because this is now sorted as the 3rd configuration in this
section/group the extra configurations for the grouped product image and
the configurable product image are now also increased.

This feature will make it possible to have this cross-sell turned off
during special sales, A/B testing or in a specific store view.