Update Invoice.php
you can't cancel invoice if state == paid, If state == paid, you can do the credit memo.

When you create an invoice and realise that it needs to be cancelled, be aware that there are some legal restrictions involved. Technically, you should never delete an issued invoice but instead use a credit note to cancel the invoice.